# Refactor bulk API to minimize allocations and increase speed.